+++ sema4ai/vscode-client/src/viewsRobots.ts
@@ -285,22 +285,13 @@ export class RobotsTreeDataProvider implements vscode.TreeDataProvider<RobotEntr
                         "tooltip": "Create the Action Package .zip file to the workspace folder",
                     },
                     {
-                        "label": "Create Package OpenAPI Spec (metadata.json)",
+                        "label": "Open the Package OpenAPI Spec (metadata.json)",
                         "uri": element.uri,
                         "robot": element.robot,
                         "iconPath": "new-file",
-                        "type": RobotEntryType.PackageMetadataToWorkspace,
+                        "type": RobotEntryType.PackageMetadata,
                         "parent": element,
-                        "tooltip": "Create metadata.json file in Action Package folder",
-                    },
-                    {
-                        "label": "View Package OpenAPI Spec (metadata.json)",
-                        "uri": element.uri,
-                        "robot": element.robot,
-                        "iconPath": "go-to-file",
-                        "type": RobotEntryType.OpenMetadataFileToWorkspace,
-                        "parent": element,
-                        "tooltip": "Open the Action Package metadata file (metadata.json)",
+                        "tooltip": "Open the Action Package OpenAPI spec (metadata.json) file",
                     },
                 ];
             } else if (element.type === RobotEntryType.Robot) {
@@ -530,17 +521,10 @@ export class RobotsTreeDataProvider implements vscode.TreeDataProvider<RobotEntr
                 "command": roboCommands.SEMA4AI_ACTION_SERVER_PACKAGE_BUILD,
             };
             treeItem.collapsibleState = vscode.TreeItemCollapsibleState.None;
-        } else if (element.type === RobotEntryType.PackageMetadataToWorkspace) {
+        } else if (element.type === RobotEntryType.PackageMetadata) {
             treeItem.command = {
-                "title": "Create Package OpenAPI Spec File to Workspace (metadata.json)",
+                "title": "Open the Package OpenAPI Spec (metadata.json)",
                 "command": roboCommands.SEMA4AI_ACTION_SERVER_PACKAGE_METADATA,
-                "arguments": [vscode.Uri.file(element.robot.directory)],
-            };
-            treeItem.collapsibleState = vscode.TreeItemCollapsibleState.None;
-        } else if (element.type === RobotEntryType.OpenMetadataFileToWorkspace) {
-            treeItem.command = {
-                "title": "Open Package OpenAPI Spec File to Workspace (metadata.json)",
-                "command": roboCommands.SEMA4AI_OPEN_ACTION_PACKAGE_METADATA,
                 "arguments": [vscode.Uri.file(element.robot.directory)],
             };
             treeItem.collapsibleState = vscode.TreeItemCollapsibleState.None;