+++ docs/changelog.md
@@ -1,9 +1,8 @@
 Unreleased
 -----------------------------
 
-- Add command and UI element to create metadata file for Action Package
+- Add command and UI element to view metadata file for Action Package
 - A prompt allowing to restart the Action Server will appear now when an environment file is modified in an Action package with active Action Server.
-- Add UI element to open the metadata file for Action Package
 - Fix build Action Package command to add selection step to select the Action Package in case there are multiple in workspace
 
 New in 2.1.0 (2024-06-25)